support switching branches on tegra-demo-distro
Add a kirkstone definition which can be used to
override branch configs on the command line, for instance:
```
kas build --update kas/swupdate-oe4t.yml:kas/include/kirkstone.yml
```

To support this, add a separate definition for the tegra-demo-distro
master branch, where we always pull the latest kas file, and the
tegra-demo-distro branch where the appropriate layer content
exists.  Disable the layers on the master branch since we only
want to use the layers from the target branch.